Character Creations and NPC Task flow

+++ MAMR Game Server/Server/MapGroupKernel/AINpc.h
@@ -131,20 +131,20 @@ public: // IRole
 	virtual DWORD	GetLev				()				{ return m_pType->GetInt(NPCTYPEDATA_LEVEL); }
 
 	virtual	DWORD	GetLife				();	//			{ return m_nCurrLife; }
-	virtual DWORD	GetMaxLife			();	//				{ return (m_AddInfo.nMaxLifeAdd + m_pType->GetInt(NPCTYPEDATA_LIFE)); }
+	virtual double	GetMaxLife			();	//				{ return (m_AddInfo.nMaxLifeAdd + m_pType->GetInt(NPCTYPEDATA_LIFE)); }
 //	virtual DWORD	GetMana				()				{ return m_nCurrMana; }
 //	virtual DWORD	GetMaxMana			()				{ return m_pType->GetInt(NPCTYPEDATA_MANA); }
 	virtual DWORD	GetMinAtk			();
 	virtual DWORD	GetMaxAtk			();
 	virtual DWORD	GetMgcMinAtk			();
 	virtual DWORD	GetMgcMaxAtk			();
-	virtual DWORD	GetAttack			()				{ return (GetMaxAtk()+GetMinAtk())/2; }	//{ return (m_pType->GetInt(NPCTYPEDATA_ATTACKMIN)+m_pType->GetInt(NPCTYPEDATA_ATTACKMAX))/2; }
+	virtual double	GetAttack			()				{ return (GetMaxAtk()+GetMinAtk())/2; }	//{ return (m_pType->GetInt(NPCTYPEDATA_ATTACKMIN)+m_pType->GetInt(NPCTYPEDATA_ATTACKMAX))/2; }
 	virtual DWORD	GetAtkHitRate		()				{ return (m_pType->GetInt(NPCTYPEDATA_ATK_HITRATE)); }
-	virtual DWORD	GetDef				();
-	virtual DWORD	GetDefence			()				{ return m_pType->GetInt(NPCTYPEDATA_DEFENCE); }
+	virtual double	GetDef				();
+	virtual double	GetDefence			()				{ return m_pType->GetInt(NPCTYPEDATA_DEFENCE); }
 	virtual DWORD	GetDefence2			()				{ return m_pType->GetInt(NPCTYPEDATA_DEFENCE2); }
 	virtual DWORD	GetDex				();
-	virtual DWORD	GetDexterity		()				{ return m_pType->GetInt(NPCTYPEDATA_DEXTERITY); }
+	virtual double	GetDexterity		()				{ return m_pType->GetInt(NPCTYPEDATA_DEXTERITY); }
 	virtual DWORD	GetDdg				();
 	virtual DWORD	GetDodge			()				{ return m_pType->GetInt(NPCTYPEDATA_DODGE); }
 //	virtual DWORD	GetMagicAtk			()				{ return 0; }
